src/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
# mysqlclient==2.1.1
packaging==21.3
pycodestyle==2.10.0
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
# sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13